pyproject.toml
[project]
name = "de1-2025-project"
version = "0.1.0"
description = ""
authors = [
    {name = "deedeetdt",email = "pongsatorn.kam@student.mahidol.edu"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "apache-flink (==1.18.0)",
    "requests (>=2.32.5,<3.0.0)",
    "kafka-python (>=2.3.0,<3.0.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"